Feature/authentication backend (#2)
* feat: added routes/controller for auth, signup and signin

* feat: added User model, controllers/routes for auth with signin/signup, sign up working and tested in postman, returning a valid jwt token + user data

* feat: signup and signin both working and tested in postman and returning a valid jwt

* feat: added middleware to check for a valid token in the request header, if truthy we will set it to req.userId

* feat: added helper route in our auth stuff to get user data

* fix: changed from adding just the id to our req object, we are now adding the whole user minus the password

* feat: get all habits now only fetches the logged in users habits based on our req.user object, all crud routes are now protected and new valid jwts

* feat: users can no longer update or delete others habits, checking the user field on the habit matches the user.id from the auth headers

+++ backend/controllers/habits.js
@@ -1,9 +1,10 @@
 const asyncHandler = require("express-async-handler");
 const Habits = require("../models/habits.js");
+const User = require("../models/user.js");
 const mongoose = require("mongoose");
 
 const getAllHabits = asyncHandler(async (req, res) => {
-  const habits = await Habits.find();
+  const habits = await Habits.find({ user: req.user.id });
   res.json(habits);
 });
 
@@ -17,6 +18,7 @@ const createHabit = asyncHandler(async (req, res) => {
     );
   }
   const newHabit = await Habits.create({
+    user: req.user.id,
     name,
     type,
     description,
@@ -34,6 +36,18 @@ const updateHabit = asyncHandler(async (req, res) => {
   }
 
   const habit = req.body;
+  const habitCheck = await Habits.findById(req.params.id);
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (habitCheck.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
 
   const updatedHabit = await Habits.findByIdAndUpdate(
     _id,
@@ -48,7 +62,18 @@ const deleteHabit = asyncHandler(async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
     throw new Error(`No habit with ID ${req.params.id} found`);
   }
+  const habit = await Habits.findById(req.params.id);
 
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (habit.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   await Habits.findByIdAndRemove(req.params.id);
   res.json({ message: `Succesfully deleted Habit with ID ${req.params.id}` });
 });
